move ReservationTable to mapf
