+++ tests/mapf/test_mapf.py
@@ -1,7 +1,8 @@
 import unittest
 from collections import defaultdict
 from w9_pathfinding import mapf
-from w9_pathfinding.envs import Graph, Grid, ReservationTable
+from w9_pathfinding.envs import Graph, Grid
+from w9_pathfinding.mapf import ReservationTable
 
 COMPLETE_ALGORITHMS = [
     {
