+++ tests/mapf/test_space_time_a_star.py
@@ -1,7 +1,7 @@
 import unittest
 from w9_pathfinding import DiagonalMovement
-from w9_pathfinding.envs import Grid, Graph, ReservationTable
-from w9_pathfinding.mapf import SpaceTimeAStar
+from w9_pathfinding.envs import Grid, Graph
+from w9_pathfinding.mapf import SpaceTimeAStar, ReservationTable
 
 
 class TestSpaceTimeAStar(unittest.TestCase):
